Improve day 16 part 2 performance by looking for non-overlapping paths
This speeds up the solution time by about 95%: 3min -> 9s.
Collects all seen paths and respective scores and then looks for
the best non-overlapping path that the agents could take.

The author of the reference below claims it takes him 800ms to run part
2, so there's likely some room to optimize here.

Also, I don't think its a bottleneck, but a lot of people seem to be
fond of Floyd's algorithm to compute the distance matrix. I guess I
could take a look at that as well.

Reference:
<https://github.com/davearussell/advent2022/blob/master/day16/solve.py>